backend/requirements.txt
anyio==4.1.0
asgiref==3.7.2
certifi==2023.11.17
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
